+++ docs/source/changelog.rst
@@ -7,6 +7,8 @@ Unreleased
 Fixed
 ~~~~~
 * The `ssbm` method will not allow non-float probabilities
+* The `Graph.weight` method will give the correct value when sets are equal
+* Prevent `sbm` method from generating duplicate edges
 
 0.3 - 2021-11-11
 ----------------